Initial parent pom and modularization
