+++ data-gen/src/main/java/io/smartcat/data/loader/Rule.java
@@ -1,0 +1,11 @@
+package io.smartcat.data.loader;
+
+public interface Rule<T> {
+
+    boolean isExclusive();
+
+    Rule<T> recalculatePrecedance(Rule<T> exclusiveRule);
+
+    T getRandomAllowedValue();
+
+}
