+++ data-gen/src/main/java/io/smartcat/data/loader/model/Address.java
@@ -1,0 +1,36 @@
+package io.smartcat.data.loader.model;
+
+/**
+ * Address dummy entity.
+ */
+public class Address {
+
+    private String city;
+    private String street;
+    private long houseNumber;
+
+    public String getCity() {
+        return city;
+    }
+
+    public void setCity(String city) {
+        this.city = city;
+    }
+
+    public String getStreet() {
+        return street;
+    }
+
+    public void setStreet(String street) {
+        this.street = street;
+    }
+
+    public long getHouseNumber() {
+        return houseNumber;
+    }
+
+    public void setHouseNumber(long houseNumber) {
+        this.houseNumber = houseNumber;
+    }
+
+}
