+++ load-gen/src/main/java/io/smartcat/App.java
@@ -1,0 +1,13 @@
+package io.smartcat;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
