--- src/main/java/io/smartcat/LoadgenApplication.java
@@ -1,15 +1,0 @@
-package io.smartcat;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * Entry point.
- */
-@SpringBootApplication
-public class LoadgenApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(LoadgenApplication.class, args);
-    }
-}
