--- src/main/java/io/smartcat/model/Address.java
@@ -1,36 +1,0 @@
-package io.smartcat.model;
-
-/**
- * Address dummy entity.
- */
-public class Address {
-
-    private String city;
-    private String street;
-    private long houseNumber;
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public String getStreet() {
-        return street;
-    }
-
-    public void setStreet(String street) {
-        this.street = street;
-    }
-
-    public long getHouseNumber() {
-        return houseNumber;
-    }
-
-    public void setHouseNumber(long houseNumber) {
-        this.houseNumber = houseNumber;
-    }
-
-}
