--- src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.data.mongodb.database=test
-spring.data.mongodb.host=localhost
-spring.data.mongodb.port=27017
