--- src/test/java/io/smartcat/model/Rule.java
@@ -1,11 +1,0 @@
-package io.smartcat.model;
-
-public interface Rule<T> {
-
-    boolean isExclusive();
-
-    Rule<T> recalculatePrecedance(Rule<T> exclusiveRule);
-
-    T getRandomAllowedValue();
-
-}
